requirements.txt
langchain>=0.0.335
langchain-openai>=0.0.5
langchain-core>=0.1.16,<0.2
pydantic>=2.5.2
fastapi>=0.109.2
uvicorn>=0.24.0
pymongo>=4.6.1
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
